refactor(wallet): types
